Solucionar Taller Nómina
* Se adiciona una validación para que no se eliminen las respuestas guardadas si la filas filas del taller de nómina vienen vacías.

+++ app/Http/Controllers/TallerController.php
@@ -864,7 +864,7 @@ class TallerController extends Controller
                     die();
                 }
             }
-            if ($respuestaTallerNomina->filasTallerNomina->isNotEmpty()) {
+            if ($respuestaTallerNomina->filasTallerNomina->isNotEmpty() && !empty($tallerNominaRespuestas)) {
                 $respuestaTallerNomina->filasTallerNomina()->delete();
             }
             foreach ($tallerNominaRespuestas as $fila) {
